Add vm modification function
Adds a function to modify an existing VM object's configuration,
either directly in-editor, or by replacing with a new XML config.
In replacement mode, there is no confirmation or diff, the new
configuration is simply written and reboot triggered if specified.
In editor mode, the users local EDITOR (or `vi`) is started with
a copy of the existing config. Upon writing and exiting, the
operator is presented with a diff of the changes and is asked to
confirm before writing them to Zookeeper.

Closes #13